Moved navigation pages to separate package
