+++ assessment/src/main/java/org/example/ui/pages/navigation/TranslatePage.java
@@ -1,5 +1,6 @@
-package org.example.ui.pages;
+package org.example.ui.pages.navigation;
 
+import org.example.ui.pages.BasePage;
 import org.openqa.selenium.WebDriver;
 
 public class TranslatePage extends BasePage {
