+++ assessment/src/test/java/org/example/ui/YandexMainTest.java
@@ -2,14 +2,14 @@ package org.example.ui;
 
 import lombok.extern.slf4j.Slf4j;
 import org.example.ui.steps.BaseSteps;
-import org.example.ui.steps.ImagesPageSteps;
+import org.example.ui.steps.navigation.ImagesPageSteps;
 import org.example.ui.steps.MainPageSteps;
-import org.example.ui.steps.MapsPageSteps;
+import org.example.ui.steps.navigation.MapsPageSteps;
 import org.example.ui.steps.market.MarketPageSteps;
-import org.example.ui.steps.MusicPageSteps;
-import org.example.ui.steps.NewsPageSteps;
-import org.example.ui.steps.TranslatePageSteps;
-import org.example.ui.steps.VideoPageSteps;
+import org.example.ui.steps.navigation.MusicPageSteps;
+import org.example.ui.steps.navigation.NewsPageSteps;
+import org.example.ui.steps.navigation.TranslatePageSteps;
+import org.example.ui.steps.navigation.VideoPageSteps;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
